refactor: reorganize message handling to use UpdateAware interface (#214)
Move view-specific message types from centralized model.go to their respective
view files, implementing the UpdateAware interface pattern for better encapsulation.

## Changes
- Moved dindContainersLoadedMsg to view_dind_process_list.go
- Moved dockerContainersLoadedMsg to view_docker_container_list.go
- Moved dockerImagesLoadedMsg to view_image_list.go
- Moved dockerNetworksLoadedMsg to view_network_list.go
- Moved dockerVolumesLoadedMsg to view_volume_list.go
- Moved projectsLoadedMsg to view_compose_project_list.go
- Moved topLoadedMsg to view_top.go
- Moved statsLoadedMsg to view_stats.go
- Moved containerFilesLoadedMsg to view_file_browser.go
- Moved fileContentLoadedMsg to view_file_content.go
- Moved inspectLoadedMsg to view_inspect.go

All view models now implement UpdateAware interface to handle their own messages,
reducing complexity in the central update.go file and improving code organization.
